fix: KrdsHelpPanel을 KrdsPanel로 변경
