--- src/components/KrdsHelpPanel/index.ts
@@ -1,8 +1,0 @@
-/**
- * KRDS HelpPanel 컴포넌트
- */
-
-import KrdsHelpPanel from './KrdsHelpPanel.vue'
-
-export { KrdsHelpPanel }
-export default KrdsHelpPanel
