+++ src/components/KrdsPanel/KrdsPanel.vue
