+++ src/components/KrdsPanel/index.ts
@@ -1,0 +1,8 @@
+/**
+ * KRDS Panel 컴포넌트
+ */
+
+import KrdsPanel from './KrdsPanel.vue'
+
+export { KrdsPanel }
+export default KrdsPanel
